Pruned redundant copy of appConfig I'd accidentally created and streamlined code per review comment
As per review feedback from Miroslav, removing safety checks that sought
to prevent Grunt crashing if .bowerrc did not exist, or did not contain
a directory entry from the mainline path and encapsulated them in a
helper function instead.  Out of sight, out of mind.

Also fixed a mis-spelling of config.js in a code comment that was irking
me.