Accounting for backwards incompatbile changes in latest Browserify.
Browserify has deprecated bundle options since this code was first
written.  There were several options previously provided to the bundle()
call that now must be passed to the browserify( ) factory method.

Also added a TODO comment about a node module that can be used to
extract source maps from a browserify bundle.  There was a pre-existing
TODO note about wanting source maps to always be generated, but only
embedded within the bundle for development builds, not production.
There seems to be no way to ask Browserify to generate its source maps
outside the bundle, but it does make reference to the Excorcist utllity
as a means to accomplish that goal all the same.