Rolling back an unnecessary experiment
I'd wondered why usemin was not using CSS concatenation.  I didn't
notice any problems with it, but I've read other people suspecting they
were having problems with it and it isn't necessary to change this
forthe patch goal to get there, but I noticed I had accidentally checked
this in instead of rolling it back, so correcting that now.

Retuning one line to iriginal state....